ci: Add crate version check (#38)
Adds a script to determine whether the local and remote versions of a
crate are the same. This script is then used in CI to determine whether
to publish a crate or not. This makes it possible to release only one of
the two published crates at a time successfully.

Signed-off-by: Dan Sover <dan.sover@avalabs.org>
Co-authored-by: Dan Sover <dan.sover@avalabs.org>
Co-authored-by: Sam Batschelet <sam.batschelet@avalabs.org>